+++ package.json
@@ -70,12 +70,13 @@
     "typescript": "^4.5.4"
   },
   "dependencies": {
+    "@ethersproject/constants": "^5.5.0",
+    "@ethersproject/logger": "^5.5.0",
     "@ethersproject/providers": "^5.5.2",
     "@harmony-js/account": "^0.1.57",
     "@harmony-js/contract": "^0.1.56",
     "@harmony-js/core": "^0.1.57",
     "@harmony-js/transaction": "^0.1.56",
-    "@harmony-js/utils": "^0.1.56",
-    "ethers": "^5.5.3"
+    "@harmony-js/utils": "^0.1.56"
   }
 }
